bug in remove from queue
git-svn-id: https://build.archivesportaleurope.net/svn/apex/trunk@858 9fd842fd-667d-44da-918d-ff91420b3ae1

+++ Dashboard/src/main/java/eu/apenet/dashboard/services/ead/EadService.java
@@ -182,7 +182,7 @@ public class EadService {
 			File file = new File(filename);
 			File aiDir = file.getParentFile();
 			ContentUtils.deleteFile(file, false);
-			if (aiDir.listFiles().length == 0){
+			if (aiDir.exists() && aiDir.listFiles().length == 0){
 				ContentUtils.deleteFile(aiDir, false);
 			}
 		}
